+++ FactureViewer/FactureViewer/ViewModel/FactureViewModel.cs
@@ -12,6 +12,6 @@ namespace FactureViewer.ViewModel
 				public string WareName { get; set; }
 				public int Count { get; set; }
 				public double Price { get; set; }
-				public int Tax { get; set; }
+				public double Tax { get; set; }
     }
 }
